+++ Tortuga.Chain/Shared/Tests/Core/MetadataTests.cs
@@ -102,9 +102,10 @@ public class MetadataTests : TestBase
 			Release(dataSource);
 		}
 	}
+
 #endif
 
-#if SQL_SERVER_MDS
+#if SQL_SERVER_MDS || SQL_SERVER_OLEDB
 
 	[DataTestMethod, BasicData(DataSourceGroup.All)]
 	public void PreloadBySchema(string dataSourceName, DataSourceType mode)
@@ -113,17 +114,33 @@ public class MetadataTests : TestBase
 		try
 		{
 			dataSource.DatabaseMetadata.Reset();
-			dataSource.DatabaseMetadata.PreloadTables("HR");
+			dataSource.DatabaseMetadata.Preload("HR");
+
 			var tables = dataSource.DatabaseMetadata.GetTablesAndViews();
-			Assert.IsTrue(tables.Any(t => t.Name.Schema == "HR"), "Expected at least one HR table");
-			Assert.IsFalse(tables.Any(t => t.Name.Schema != "HR"), "Expected no non-HR tables");
+			Assert.IsTrue(tables.Any(t => t.IsTable && t.Name.Schema == "HR"), "Expected at least one HR table");
+			Assert.IsFalse(tables.Any(t => t.IsTable && t.Name.Schema != "HR"), "Expected no non-HR tables");
+
+			var views = dataSource.DatabaseMetadata.GetTablesAndViews();
+			Assert.IsTrue(views.Any(t => !t.IsTable && t.Name.Schema == "HR"), "Expected at least one HR view");
+			Assert.IsFalse(views.Any(t => !t.IsTable && t.Name.Schema != "HR"), "Expected no non-HR views");
+
+			var funcs = dataSource.DatabaseMetadata.GetStoredProcedures();
+			Assert.IsTrue(funcs.Any(t => t.Name.Schema == "HR"), "Expected at least one HR func");
+			Assert.IsFalse(funcs.Any(t => t.Name.Schema != "HR"), "Expected no non-HR funcs");
+
+			var procs = dataSource.DatabaseMetadata.GetStoredProcedures();
+			Assert.IsTrue(procs.Any(t => t.Name.Schema == "HR"), "Expected at least one HR proc");
+			Assert.IsFalse(procs.Any(t => t.Name.Schema != "HR"), "Expected no non-HR procs");
+
+			var types = dataSource.DatabaseMetadata.GetUserDefinedTableTypes();
+			Assert.IsTrue(types.Any(t => t.Name.Schema == "HR"), "Expected at least one HR type");
+			Assert.IsFalse(types.Any(t => t.Name.Schema != "HR"), "Expected no non-HR types");
 		}
 		finally
 		{
 			Release(dataSource);
 		}
 	}
-
 
 #endif